mv job stream / job list
